config/requirements.txt
asgiref==3.9.1
bleach==6.2.0
crispy-bootstrap5==2025.6
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-crispy-forms==2.4
django-csp==4.0
django-environ==0.12.0
django-filter==23.3
django-heroku==0.3.1
django-taggit==6.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
joblib==1.5.1
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.3.0
packaging==25.0
pandas==2.3.0
Pillow==10.1.0
psycopg2==2.9.11
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
webencodings==0.5.1
whitenoise==6.6.0
